% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_nz_data}
\alias{get_nz_data}
\title{Get Dataset}
\usage{
get_nz_data(resource_id)
}
\arguments{
\item{resource_id}{A string}
}
\value{
A dataframe
}
\description{
'get_nz_data()' returns a dataframe from the resource identified in resource_id.
}
\examples{
get_nz_data('4b292323-9fcc-41f8-814b-3c7b19cf14b3')
}
