% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(resource_id)
}
\arguments{
\item{resource_id}{A string}
}
\value{
large list
}
\description{
Load data
}
\examples{
streets <- '2f4e30a7-091f-4aed-b18f-4b21cc19e6b4'
load_data(streets)
}
