% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{nz_dataset_metadata}
\alias{nz_dataset_metadata}
\title{Get Dataset Metadata}
\usage{
nz_dataset_metadata(dataset_id)
}
\arguments{
\item{dataset_id}{A string}
}
\value{
A dataframe
}
\description{
'nz_dataset_metadata()' returns the metadata for the dataset identified by 'dataset_id'
}
\examples{
nz_dataset_metadata('directory-of-educational-institutions')
}
