% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_metadata.R
\name{nz_resource_metadata}
\alias{nz_resource_metadata}
\title{Get Resource Metadata}
\usage{
nz_resource_metadata(resource_id)
}
\arguments{
\item{resource_id}{A string}
}
\value{
A dataframe
}
\description{
'nz_resource_metadata()' returns the metadata for the dataset identified by 'resource_id'
}
\examples{
nz_resource_metadata('4b292323-9fcc-41f8-814b-3c7b19cf14b3')
}
