% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzdata-package.R
\docType{package}
\name{nzdata-package}
\alias{nzdata}
\alias{nzdata-package}
\title{nzdata: API Wrapper for data.govt.nz}
\description{
A wrapper for the API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/courtneyruss/nzdata}
  \item Report bugs at \url{https://github.com/courtneyruss/nzdata/issues}
}

}
\author{
\strong{Maintainer}: Courtney Russ \email{courtney@courtneyruss.com} (\href{https://orcid.org/0009-0009-1278-2808}{ORCID})

}
\keyword{internal}
